requirements.txt
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
future==0.18.2
idna==2.8
Jinja2==2.10.3
MarkupSafe==1.1.1
peewee==3.11.2
pycparser==2.19
python-telegram-bot==12.2.0
requests==2.22.0
six==1.13.0
tornado==6.0.3
urllib3==1.25.7
